pyproject.toml
[tool.poetry]
name = "tuna"
version = "0.1.0"
description = ""
authors = ["ElderMillenial <eldermillenial@protonmail.com>"]
readme = "README.md"
packages = [{include = "tuna", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pycardano = "^0.11.1"
blockfrost-python = "^0.5.3"
setuptools = "^71.1.0"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
